Calendar Render Update
Changed how calendar renders. Used to rely on single DayCard for both weekly and monthly views, causing constraints on building the toggleable views.

This commit creates a DayCard for weeks and a MonthDayCard to modularize the the rendering. There are two calendar view options, WeekView and MonthView
In weekly mode, the first two event names of each day are displayed with a counter of how many more there are that day available. If no events, displays, "No SHPE :("
In month view, event names are not displayed and only the card color changes if there are events.
In both, the users are able to select a date and see further event information;.

HomeViewModel Update:
should now be able to fetch events for specific days and has the ability to change weeks
one more function that finds if an event happens on a specific date is declared.