State Requirement Change
Changed the app to use a separate state file for the HomeViewModel and UI.

HomeScreenState.kt has the new data class used for state

HomeViewModel.kt is updated to used the new state using MutableStateFlow.

HomeScreen.kt: Some functions are updated to use the new homeState variable from the viewModel.