Update to load more events at a time
HomeScreen.kt:
Is updated to have EventCardFeedSimple that will just load events for the next month when the user reaches the end of the list. If no events, will just try to reload that same month to prevent the loading of unnecessary events.

HomeScreenState.kt:
Updated to include the months that are loaded and a variable that allows us to tell if the month is loading or months are requested or empty.

HomeViewModel.kt:
Updated to have getDaysInMonth instead of in the UI, also changes the fetchCalendarEvents to update the month that was requested.