Updated the app to have repository for notification settings
MainActivity.kt:
Initializes the notification repository and passes to the HomeViewModel, which is passed to the HomeScreen.

HomeScreen.kt:
updates the previews to work with the new repository need in in the view model

HomeViewModel.kt:
includes the HomeViewModelFactory to be able to initialize the HomeViewModel with the repo requirement.