Update to how events are loaded to prevent duplication
In the HomeViewModel, the fetching of events from the server is from today to 4 months after.

The load events now will check database first and then load events from the server if there are none. However, when the app is launched again, it does not automatically load from the server, but another state could be used such as just launched.

There are minor changes to names of start and end dates.

HomeScreenState was updated to include a month being displayed.

The home screen has an extra testing button to load events from the server. This is used as when the app is launched, it does not reload automatically from the server.

The home screen now ensures that the list of events do not duplicate events when storing. Filtering per day is done slightly different. There is no long a scrolling to the bottom to load more dynamically, rather a set span of now to 4 months from now are loaded.

In the Event db, the events are unique by the ID given by google calendar.