UI Update 3
HomeScreen: Created a pull to refresh feature to enhance usability. Added a pull to refresh lazy column (from vid) that could be used in other places. The date now changes depending on the month of which the currently displayed events are in.

HomeViewModel: Added the pull to refresh function and removed unncessary functions.