Notifications Update
The app now initializes a notifications manager to deal with context alone.

The view model is mainly reorganized. There is an additional function to determine if notifications are supposed to be enabled for the type. Used in the fetch events.

Notifications util: updated to be an object rather than a class, allows for context to be dealt with individually, Fixed the issue with overwriting notifications. There are now individual notifications and notification grouping for multiple notifications.

Enabling all notifications now allows for notifications to be enabled.

Currently, cancelling event notifications need to be implemented. Other than that, I believe that notifications are done. Then paging should be it and setting all of the pages together.